Users added to table and sorted
This commit takes the users from the ajax command, places them into user
objects, then exports the user data to the html table.  This also allows
for a custom sort function to be injected so that in the future we can
sort by column by clicking on the column headers.